server/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyYAML==6.0.1
Werkzeug==2.3.7
mysqlclient==2.2.0
PyMySQL==1.1.0